Separate the different schemas for the database. Create script to set the environment to work locally. Edit and improve the Makefile adding clean and support for environment creation.
