+++ Dockerfile.test
@@ -5,5 +5,6 @@ RUN pip install -r /tmp/requirements-test.txt
 
 RUN mkdir /app
 COPY ./tests/ /app/backend/tests/
+COPY ./src/ /app/backend/src/
 
 WORKDIR /app/backend/tests
