+++ configs/setenv.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+python3 -m venv venv
+source ./venv/bin/activate
+python3 -m pip install -r ./configs/requirements-dev.txt
