+++ src/db/__init__.py
@@ -1,7 +1,8 @@
 from pony.orm import Database
 
 db = Database()
-from schemas.room import Room
+from schemas.room import Room, User
+from schemas.game import Game, Player, Card, Deck
 
 # Conectamos el objeto `db` con la base de dato.
 db.bind('sqlite', 'example.sqlite', create_db=True)
