+++ src/schemas/enums.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class Role(Enum):
+    human = 0
+    thing = 1
+    infected = 2
