+++ src/schemas/game.py
@@ -1,23 +1,8 @@
-from pony.orm import *
+from pony.orm import (PrimaryKey, Required, Set, Optional)
+from .room import Room, User
+from db import db
 
-db = Database()
-
-
-class User(db.Entity):
-    id = PrimaryKey(int, auto=True, unsigned=True)
-    name = Required(str, 30)
-    id_lobby = Optional('Lobby')
-
-
-class Lobby(db.Entity):
-    id = PrimaryKey(int, auto=True, unsigned=True)
-    name = Required(str, 30)
-    id_user = Set(User)
-    minimum_cnt_users = Required(int, default=4, unsigned=True)
-    maximum_cnt_users = Required(int, default=12)
-
-
-class Game(Lobby):
+class Game(Room):
     round_left_direction = Required(bool, default=0)
     actual_phase = Required(str, default='Draw')  # Draw, Play, Discard
     actual_position = Optional(int, default=1, unsigned=True)
@@ -29,7 +14,6 @@ class Player(User):
     round_position = Required(int, unique=True, unsigned=True)
     alive = Required(bool, default=1)
 
-
 class Card(db.Entity):
     id = PrimaryKey(int, auto=True, unsigned=True)
     name = Required(str, 30)
@@ -37,13 +21,9 @@ class Card(db.Entity):
     type = Required(str, default='Action')  # Action, Defense, Infection, Obstacle, Panic
     decks = Set('Deck')
 
-
 class Deck(db.Entity):
     games = Required(Game)
     cards = Required(Card)
     is_available_deck = Required(bool)
     PrimaryKey(games, cards)
 
-
-
-db.generate_mapping()
